+++ CHANGELOG.md
@@ -9,9 +9,9 @@
 - Bump WorkManager to v2.5.0
 - Fix the Room query thread pool executor to run all available threads
 - Load list of `colonyOrVillage` from the `PatientAddress` table
-- Stop loading all patients after current facility is loaded in `InstantSearchScreen`
 - Change `TeleconsultRecordScreen` to a fragment
 - Change `EditMedicinesScreen` to a fragment
+- Only load the alphabetical patient log once when the screen is opened
 - [In Progress: 20 Jan 2021] Material Theme-ing Migration
 - [In Progress: 08 Feb 2021] Migrate app to use ViewBinding
 
@@ -20,6 +20,7 @@
 - Show progress when linking bp passport to patient
 - Show progress when saving medical history for new patient
 - Change prescriptions end icon to a chevron icon instead of dropdown
+- Stop loading patients from other facilities in the alphabetical patient log
 
 ## 2021-02-26-7647
 ### Internal